requirements.txt
aiohttp==3.12.15
prometheus-client==0.22.1
pyyaml==6.0.2